# go/models/model_job.go
/*
 * DistributedCronScheduler
 *
 * A simple distributed job scheduler running on top of AWS
 *
 * API version: 1.0.0
 * Contact: apollo-anonymous+demo@amazon.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package cronscheduler

type Job struct {

	JobId int64 `json:"jobId"`

	JobName string `json:"jobName"`

	// cron expression
	JobSchedule string `json:"jobSchedule"`

	// job execution target arn
	JobExecutionTargetArn string `json:"jobExecutionTargetArn"`

	// job status
	Status string `json:"status,omitempty"`
}